requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.5
djangorestframework==3.10.3
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
numpy==1.16.4
pandas==0.25.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
rest-pandas==1.1.0
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0
wincertstore==0.2
wrapt==1.11.2
